Soundness and completeness of nbe for STLCL
Following the lesson learned from proving the completeness of our
nbe for stlc:

+ dropping fancy models: coercions and use of the upward closure
of the model get in the way when proving completeness. The simple/
model is quite simple and the normalization algorithm is cristal
clear.

+ soundness and completeness are proved separately using logical
relations ("reduces to" for soundness and a more involved "uniform
and extensional" pair of mutually defined ones for completeness).

Nb: The completeness proof is heavily based on C. Coquand's paper
on nbe for a stlc with explicit substitution