using nbe to build a tactic in Agda
Given that our equational theory is decidable and compatible
with Agda's one (if we assume functional extensionality), we
can build a solver using our nbe in order to prove lemmas in
Agda.

Here is the machinery making the link between the terms we
identify with our equational theory and the theorems that
hold in Agda.
The next step is to use the quoting mechanism (well-described
in e.g. Paul van der Walt's Master's thesis.